Cambia nombre de Packete Sample a GUI. Depura Codigo. Falta Trabajar en Ventanas
