--- bin/amss/sample/Views/perfil.fxml
@@ -1,129 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.image.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="547.0" prefWidth="795.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="amss.sample.Controllers.Perfil_Controller">
-   <children>
-      <VBox prefHeight="547.0" prefWidth="794.0">
-         <children>
-            <MenuBar>
-              <menus>
-                <Menu mnemonicParsing="false" onAction="#transition_Back" onMenuValidation="#transition_Back" onShowing="#transition_Back" onShown="#transition_Back" text="Regresar">
-                     <items>
-                        <CustomMenuItem mnemonicParsing="false" onAction="#transition_Back" text="Regresar">
-                          <content>
-                            <CheckBox text="Click Me" />
-                          </content>
-                        </CustomMenuItem>
-                     </items>
-                </Menu>
-              </menus>
-            </MenuBar>
-            <HBox prefHeight="519.0" prefWidth="794.0">
-               <children>
-                  <VBox prefHeight="519.0" prefWidth="208.0">
-                     <children>
-                        <ImageView fitHeight="195.0" fitWidth="237.0" pickOnBounds="true" preserveRatio="true">
-                           <image>
-                              <Image url="@../../../old_user.png" />
-                           </image>
-                        </ImageView>
-                        <Separator orientation="VERTICAL" prefHeight="26.0" prefWidth="2.0" />
-                        <Label text="Informacion General" />
-                        <Button mnemonicParsing="false" onAction="#loadInfo" text="Load" />
-                     </children>
-                  </VBox>
-                  <TabPane prefHeight="519.0" prefWidth="593.0" tabClosingPolicy="UNAVAILABLE">
-                    <tabs>
-                      <Tab text="Pastillero">
-                        <content>
-                          <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
-                                 <children>
-                                    <TableView layoutX="-1.0" layoutY="-1.0" prefHeight="486.0" prefWidth="592.0">
-                                      <columns>
-                                        <TableColumn prefWidth="163.4298858642578" text="Medicina" />
-                                        <TableColumn minWidth="0.0" prefWidth="171.5269775390625" text="Horario" />
-                                          <TableColumn prefWidth="250.38726806640625">
-                                             <graphic>
-                                                <Button mnemonicParsing="false" prefHeight="27.0" prefWidth="76.0" text="Agregar" textAlignment="CENTER" />
-                                             </graphic>
-                                          </TableColumn>
-                                      </columns>
-                                    </TableView>
-                                 </children>
-                              </AnchorPane>
-                        </content>
-                      </Tab>
-                        <Tab text="Medicinas">
-                           <content>
-                              <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
-                                 <children>
-                                    <TableView layoutX="-1.0" layoutY="-1.0" prefHeight="486.0" prefWidth="592.0">
-                                       <columns>
-                                          <TableColumn prefWidth="106.41929626464844" text="Nombre" />
-                                          <TableColumn minWidth="0.0" prefWidth="277.7420654296875" text="Cantidad" />
-                                          <TableColumn prefWidth="196.25799560546875">
-                                             <graphic>
-                                                <Button mnemonicParsing="false" prefHeight="27.0" prefWidth="80.0" text="Agregar" />
-                                             </graphic>
-                                          </TableColumn>
-                                       </columns>
-                                    </TableView>
-                                 </children>
-                              </AnchorPane>
-                           </content>
-                        </Tab>
-                        <Tab text="Familiares">
-                           <content>
-                              <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
-                                 <children>
-                                    <TableView layoutX="-1.0" layoutY="-1.0" prefHeight="486.0" prefWidth="592.0">
-                                       <columns>
-                                          <TableColumn prefWidth="76.34140014648438" text="Nombre" />
-                                          <TableColumn prefWidth="76.68292236328125" text="Telefono" />
-                                          <TableColumn prefWidth="163.4298858642578">
-                                             <graphic>
-                                                <Button mnemonicParsing="false" prefHeight="27.0" prefWidth="80.0" text="Agregar" />
-                                             </graphic>
-                                          </TableColumn>
-                                       </columns>
-                                    </TableView>
-                                 </children>
-                              </AnchorPane>
-                           </content>
-                        </Tab>
-                        <Tab text="Recetas">
-                           <content>
-                              <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
-                                 <children>
-                                    <TableView layoutX="-1.0" layoutY="-1.0" prefHeight="486.0" prefWidth="592.0">
-                                       <columns>
-                                          <TableColumn prefWidth="175.06434631347656" text="Medicina" />
-                                          <TableColumn prefWidth="266.9357604980469" text="Fecha Inicio" />
-                                          <TableColumn minWidth="0.0" prefWidth="143.527099609375">
-                                             <graphic>
-                                                <Button mnemonicParsing="false" onAction="#transition_RecetaForm" prefHeight="27.0" prefWidth="80.0" text="Agregar" />
-                                             </graphic>
-                                          </TableColumn>
-                                       </columns>
-                                    </TableView>
-                                 </children>
-                              </AnchorPane>
-                           </content>
-                        </Tab>
-                    </tabs>
-                  </TabPane>
-               </children>
-            </HBox>
-         </children>
-      </VBox>
-   </children>
-</Pane>
