--- bin/amss/sample/Views/recetaForm.fxml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.image.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="547.0" prefWidth="795.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="amss.sample.Controllers.RecetaForm_Controller">
-   <children>
-      <VBox prefHeight="547.0" prefWidth="794.0">
-         <children>
-            <MenuBar>
-               <menus>
-                  <Menu mnemonicParsing="false" onAction="#transition_Back" onMenuValidation="#transition_Back" onShowing="#transition_Back" onShown="#transition_Back" text="Regresar">
-                     <items>
-                        <CustomMenuItem mnemonicParsing="false" onAction="#transition_Back" text="Regresar">
-                           <content>
-                              <CheckBox text="Click Me" />
-                           </content>
-                        </CustomMenuItem>
-                     </items>
-                  </Menu>
-               </menus>
-            </MenuBar>
-            <Label prefHeight="70.0" prefWidth="794.0" text="                                                             Nueva Receta para ______             " textAlignment="CENTER">
-               <font>
-                  <Font size="15.0" />
-               </font>
-            </Label>
-            <HBox prefHeight="56.0" prefWidth="794.0">
-               <children>
-                  <VBox prefHeight="483.0" prefWidth="143.0">
-                     <children>
-                        <Label prefHeight="37.0" prefWidth="158.0" text="Nombre Medicamento:" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <Label prefHeight="37.0" prefWidth="158.0" text="Frecuecia" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <Label prefHeight="37.0" prefWidth="158.0" text="Dosis" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <Label prefHeight="37.0" prefWidth="158.0" text="Cantidad" />
-                     </children>
-                  </VBox>
-                  <VBox prefHeight="483.0" prefWidth="492.0">
-                     <children>
-                        <TextField prefHeight="36.0" prefWidth="498.0" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <TextField layoutX="10.0" layoutY="10.0" prefHeight="36.0" prefWidth="498.0" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <TextField layoutX="10.0" layoutY="46.0" prefHeight="36.0" prefWidth="498.0" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <TextField layoutX="10.0" layoutY="82.0" prefHeight="36.0" prefWidth="498.0" />
-                     </children>
-                  </VBox>
-               </children>
-            </HBox>
-            <HBox prefHeight="61.0" prefWidth="608.0">
-               <children>
-                  <Separator prefHeight="0.0" prefWidth="349.0" />
-                  <Button mnemonicParsing="false" text="Button" textAlignment="CENTER" />
-               </children>
-            </HBox>
-         </children>
-      </VBox>
-   </children>
-</Pane>
