+++ src/amss/GUI/Views/staff.fxml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.image.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.text.*?>
+<?import java.lang.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="547.0" prefWidth="795.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="amss.GUI.Controllers.Inquilinos_Controller">
+   <children>
+      <VBox prefHeight="547.0" prefWidth="794.0">
+         <children>
+            <MenuBar>
+               <menus>
+                  <Menu mnemonicParsing="false" onAction="#transition_Back" onMenuValidation="#transition_Back" onShowing="#transition_Back" onShown="#transition_Back" text="Regresar">
+                     <items>
+                        <CustomMenuItem mnemonicParsing="false" onAction="#transition_Back" text="Regresar">
+                           <content>
+                              <CheckBox text="Click Me" />
+                           </content>
+                        </CustomMenuItem>
+                     </items>
+                  </Menu>
+               </menus>
+            </MenuBar>
+            <ToolBar prefHeight="21.0" prefWidth="600.0">
+               <items>
+                  <Label text="Nombre:" />
+                  <TextField prefHeight="27.0" prefWidth="205.0" />
+                  <Separator prefHeight="24.0" prefWidth="231.0" />
+                  <Label layoutX="10.0" layoutY="13.0" text="Filtros:" />
+                  <ChoiceBox prefHeight="5.0" prefWidth="130.0" />
+                  <Button mnemonicParsing="false" onAction="#transition_Back" prefHeight="19.0" prefWidth="68.0" text="Buscar" />
+               </items>
+            </ToolBar>
+            <TableView onMouseClicked="#transition_Perfil" prefHeight="479.0" prefWidth="794.0">
+               <columns>
+                  <TableColumn prefWidth="218.89601337909698" text="Imagen" />
+                  <TableColumn minWidth="7.10284423828125" prefWidth="128.31939697265625" text="Turno" />
+                  <TableColumn prefWidth="129.65447998046875" text="Atributo" />
+               </columns>
+            </TableView>
+         </children>
+      </VBox>
+   </children>
+</Pane>
