--- src/amss/sample/Views/emergenciaForm.fxml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.image.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="547.0" prefWidth="795.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="amss.sample.Controllers.InquilinoForm_Controller">
-   <children>
-      <VBox prefHeight="547.0" prefWidth="794.0">
-         <children>
-            <MenuBar>
-               <menus>
-                  <Menu mnemonicParsing="false" onAction="#transition_Back" onMenuValidation="#transition_Back" onShowing="#transition_Back" onShown="#transition_Back" text="Regresar">
-                     <items>
-                        <CustomMenuItem mnemonicParsing="false" onAction="#transition_Back" text="Regresar">
-                           <content>
-                              <CheckBox text="Click Me" />
-                           </content>
-                        </CustomMenuItem>
-                     </items>
-                  </Menu>
-               </menus>
-            </MenuBar>
-            <Label prefHeight="31.0" prefWidth="794.0" text="                                                                            Nueva Eventualidad " textAlignment="CENTER">
-               <font>
-                  <Font size="15.0" />
-               </font>
-            </Label>
-            <Separator opacity="0.0" orientation="VERTICAL" prefHeight="29.0" prefWidth="7.0" />
-            <HBox prefHeight="344.0" prefWidth="794.0">
-               <children>
-                  <VBox prefHeight="324.0" prefWidth="143.0">
-                     <children>
-                        <Label prefHeight="37.0" prefWidth="158.0" text="                   Nombre:" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <Label prefHeight="37.0" prefWidth="158.0" text="              Descripcion:" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="201.0" prefWidth="16.0" />
-                        <Label layoutX="10.0" layoutY="69.0" prefHeight="37.0" prefWidth="158.0" text="                   Servicio:" />
-                     </children>
-                  </VBox>
-                  <VBox prefHeight="483.0" prefWidth="492.0">
-                     <children>
-                        <ChoiceBox prefHeight="27.0" prefWidth="493.0" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="22.0" prefWidth="3.0" />
-                        <TextArea prefHeight="200.0" prefWidth="200.0" />
-                        <Separator opacity="0.0" orientation="VERTICAL" prefHeight="52.0" prefWidth="9.0" />
-                        <TextField />
-                     </children>
-                  </VBox>
-               </children>
-            </HBox>
-            <Separator opacity="0.0" orientation="VERTICAL" prefHeight="43.0" prefWidth="7.0" />
-            <HBox prefHeight="61.0" prefWidth="608.0">
-               <children>
-                  <Separator opacity="0.0" prefHeight="0.0" prefWidth="349.0" />
-                  <Button mnemonicParsing="false" text="Agregar" textAlignment="CENTER" />
-               </children>
-            </HBox>
-         </children>
-      </VBox>
-   </children>
-</Pane>
