--- src/amss/sample/Views/inquilinos.fxml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.image.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="547.0" prefWidth="795.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="amss.sample.Controllers.Inquilinos_Controller">
-   <children>
-      <VBox prefHeight="547.0" prefWidth="794.0">
-         <children>
-            <MenuBar>
-              <menus>
-                <Menu mnemonicParsing="false" onAction="#transition_Back" onMenuValidation="#transition_Back" onShowing="#transition_Back" onShown="#transition_Back" text="Regresar">
-                     <items>
-                        <CustomMenuItem mnemonicParsing="false" onAction="#transition_Back" text="Regresar">
-                          <content>
-                            <CheckBox text="Click Me" />
-                          </content>
-                        </CustomMenuItem>
-                     </items>
-                </Menu>
-              </menus>
-            </MenuBar>
-            <ToolBar prefHeight="21.0" prefWidth="600.0">
-              <items>
-                  <Label text="Nombre:" />
-                  <TextField prefHeight="27.0" prefWidth="205.0" />
-                  <Separator prefHeight="24.0" prefWidth="231.0" />
-                  <Label layoutX="10.0" layoutY="13.0" text="Filtros:" />
-                  <ChoiceBox prefHeight="5.0" prefWidth="130.0" />
-                <Button mnemonicParsing="false" prefHeight="19.0" prefWidth="68.0" text="Buscar" />
-                  <Button mnemonicParsing="false" onAction="#transition_NuevoInquilino" text="+" />
-              </items>
-            </ToolBar>
-            <TableView fx:id="inquilinosTable" onMouseClicked="#select_Inquilino" prefHeight="479.0" prefWidth="794.0">
-              <columns>
-                <TableColumn fx:id="Uuid" prefWidth="218.89601337909698" text="Id" />
-                <TableColumn fx:id="Nombre" prefWidth="142.32769775390625" text="Nombre" />
-                  <TableColumn fx:id="Edad" prefWidth="88.67227172851562" text="Edad" />
-                  <TableColumn fx:id="Cuarto" prefWidth="90.6805419921875" text="Cuarto" />
-                  <TableColumn fx:id="Direccion" minWidth="7.10284423828125" prefWidth="128.31939697265625" text="Direccion" />
-              </columns>
-            </TableView>
-         </children>
-      </VBox>
-   </children>
-</Pane>
